refactor: extract SupportDataTypes.swift from SupportDataController.swift
Extracts 5 support-data types (3 result/completion structs + 1 failure enum
+ SupportDataActionPresenter class) from SupportDataController.swift into a
new SupportDataTypes.swift.

Byte-identical move; SHA-256 51c5e7076b18bcdb190f86d21ffff448404794427eff70636d481607d1975469
verified against origin/main:SupportDataController.swift lines 3-133.

Applied memory rule: verified @MainActor at line 134 (for SupportDataController
class) is preserved before deleting the type block 3-133.

Reduces SupportDataController.swift from 281 to 150 lines.

Tests: 667 (unchanged).
